feat(settings): 사용자 메모리(프로필)를 별도 '내 정보' 화면으로 분리
설정 탭에 프로필(나의 리듬)을 인라인으로 펼쳐두던 것을, 설정에서 '내 정보'로 진입하는
전용 화면(MyInfoScreen)으로 옮겼다. 설정은 링크 한 줄만 남기고, 내 정보 화면에서
계정·리듬/선호를 조회하고 그대로 편집(#A-2)한다. 편집값은 재인터뷰 시드에 우선 반영.

- 신규 MyInfoScreen: 계정 카드 + 프로필(집중 시간대·집중 길이·회복 톤·알림 빈도·
  최소 단위·휴식 수용) 조회/편집. 프로필 없으면 안내 empty-state.
- ScreenId 에 'my-info' 추가, NAV_META(back: settings) + 렌더 등록.
- SettingsScreen: 프로필 섹션 제거 → '내 정보' 진입 행. 이동한 헬퍼 컴포넌트/미사용 import 정리.
